# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Braintam EdTech API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recordings:
    get:
      operationId: listRecordings
      tags: [recordings]
      summary: List recorded classes
      parameters:
        - name: grade
          in: query
          schema:
            type: integer
        - name: subjectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recording"
  /recordings/{id}:
    get:
      operationId: getRecording
      tags: [recordings]
      summary: Get recording detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"

  # ANIMATED VIDEOS
...
  /tests:
    get:
      operationId: listTests
      tags: [tests]
      summary: List tests and quizzes
      parameters:
        - name: grade
          in: query
          schema:
            type: integer
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, ongoing, completed]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Test"
  /tests/{id}:
    get:
      operationId: getTest
      tags: [tests]
      summary: Get test detail with questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestDetail"
  /tests/{id}/submit:
    post:
      operationId: submitTest
      tags: [tests]
      summary: Submit test answers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestSubmission"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"

  # STUDENT DASHBOARD & PROFILE
...
  /student/profile:
    get:
      operationId: getStudentProfile
      tags: [student]
      summary: Get student profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
    patch:
      operationId: updateStudentProfile
      tags: [student]
      summary: Update student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
...
  /student/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [student]
      summary: Get grade leaderboard
      parameters:
        - name: grade
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
components:
  schemas:
...
    Recording:
      type: object
      required: [id, title, subjectId, subjectName, grade, recordedAt, teacher, videoUrl, duration]
      properties:
        id:
          type: integer
        title:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
        grade:
          type: integer
        recordedAt:
          type: string
        teacher:
          type: string
        videoUrl:
          type: string
        duration:
          type: integer
        thumbnailUrl:
          type: ["string", "null"]
        views:
          type: ["integer", "null"]

    # ANIMATED VIDEO
...
    Test:
      type: object
      required: [id, title, subjectId, subjectName, grade, scheduledAt, duration, totalQuestions, status]
      properties:
        id:
          type: integer
        title:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
        grade:
          type: integer
        scheduledAt:
          type: string
        duration:
          type: integer
        totalQuestions:
          type: integer
        status:
          type: string
          enum: [upcoming, ongoing, completed]
        score:
          type: ["number", "null"]
        maxScore:
          type: ["number", "null"]
    TestDetail:
      type: object
      required: [id, title, subjectId, grade, duration, questions]
      properties:
        id:
          type: integer
        title:
          type: string
        subjectId:
          type: integer
        grade:
          type: integer
        duration:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
    Question:
      type: object
      required: [id, text, options, order]
      properties:
        id:
          type: integer
        text:
          type: string
        options:
          type: array
          items:
            type: string
        order:
          type: integer
        imageUrl:
          type: ["string", "null"]
    TestSubmission:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/TestAnswer"
    TestAnswer:
      type: object
      required: [questionId, selectedOption]
      properties:
        questionId:
          type: integer
        selectedOption:
          type: integer
    TestResult:
      type: object
      required: [score, maxScore, percentage, passed]
      properties:
        score:
          type: number
        maxScore:
          type: number
        percentage:
          type: number
        passed:
          type: boolean
        correctAnswers:
          type: integer
        wrongAnswers:
          type: integer

    # STUDENT
    StudentProfile:
      type: object
      required: [id, name, grade, points]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        grade:
          type: integer
        avatarUrl:
          type: ["string", "null"]
        points:
          type: integer
        rank:
          type: ["integer", "null"]
        school:
          type: ["string", "null"]
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        school:
          type: string
        avatarUrl:
          type: string
...
    LeaderboardEntry:
      type: object
      required: [rank, studentName, points, grade]
      properties:
        rank:
          type: integer
        studentName:
          type: string
        points:
          type: integer
        grade:
          type: integer
        avatarUrl:
          type: ["string", "null"]
        school:
          type: ["string", "null"]