# lib/api-zod/src/generated/types/testSubmission.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Braintam EdTech API
 * OpenAPI spec version: 0.1.0
 */
import type { TestAnswer } from './testAnswer';

export interface TestSubmission {
  answers: TestAnswer[];
}